requirements.txt
pyyaml>=6.0
pydantic>=2.3.0
paho-mqtt>=1.6.1
pyproj